chore: rename group edit modal
